+++ resources/hpxml-measures/HPXMLtoOpenStudio/tests/test_enclosure.rb
@@ -667,6 +667,28 @@ class HPXMLtoOpenStudioEnclosureTest < MiniTest::Test
     assert_in_delta(0.247, a_ext_ratio, 0.001)
   end
 
+  def test_aspect_ratios
+    # Test single-family attached
+    hpxml = _create_hpxml('base-bldgtype-single-family-attached.xml')
+    wall_outside = hpxml.walls.select { |w| w.exterior_adjacent_to == HPXML::LocationOutside && w.interior_adjacent_to == HPXML::LocationLivingSpace }[0]
+    wall_other_housing_unit = hpxml.walls.select { |w| w.exterior_adjacent_to == HPXML::LocationOtherHousingUnit && w.interior_adjacent_to == HPXML::LocationLivingSpace }[0]
+
+    wall_height = hpxml.building_construction.average_ceiling_height
+    left_right_wall_length = wall_other_housing_unit.area / wall_height
+    front_back_wall_length = ((wall_outside.area / wall_height) - left_right_wall_length) / 2.0
+    assert_in_delta(0.6667, front_back_wall_length / left_right_wall_length, 0.01)
+
+    # Test multifamily
+    hpxml = _create_hpxml('base-bldgtype-multifamily.xml')
+    wall_outside = hpxml.walls.select { |w| w.exterior_adjacent_to == HPXML::LocationOutside && w.interior_adjacent_to == HPXML::LocationLivingSpace }[0]
+    wall_other_housing_unit = hpxml.walls.select { |w| w.exterior_adjacent_to == HPXML::LocationOtherHousingUnit && w.interior_adjacent_to == HPXML::LocationLivingSpace }[0]
+
+    wall_height = hpxml.building_construction.average_ceiling_height
+    left_right_wall_length = wall_other_housing_unit.area / wall_height
+    front_back_wall_length = ((wall_outside.area / wall_height) - left_right_wall_length) / 2.0
+    assert_in_delta(0.6667, front_back_wall_length / left_right_wall_length, 0.01)
+  end
+
   def _check_surface(hpxml_surface, os_surface, expected_layer_names)
     os_construction = os_surface.construction.get.to_LayeredConstruction.get
 
@@ -788,7 +810,7 @@ class HPXMLtoOpenStudioEnclosureTest < MiniTest::Test
     model = OpenStudio::Model::Model.new
 
     # get arguments
-    args_hash['output_dir'] = 'tests'
+    args_hash['output_dir'] = File.dirname(__FILE__)
     arguments = measure.arguments(model)
     argument_map = OpenStudio::Measure.convertOSArgumentVectorToMap(arguments)