+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Fixes
 - Reduce housing characteristic file size by relaxing the six digit float format in the housing characteristics ([#877](https://github.com/NREL/resstock/pull/877))
 - Fix minor bug in sampling probability calculation ([#934](https://github.com/NREL/resstock/pull/934)) 
 - Rename sources subfolders so all tsv_makers can be imported as packages ([#959](https://github.com/NREL/resstock/pull/959))
+- Fix heating and cooling auto-season inputs ([#975](https://github.com/NREL/resstock/pull/975))
+- Remove Void from dependency columns in TSVs and update tests.([#981](https://github.com/NREL/resstock/pull/981))
 
 ## ResStock v2.5.0
 ###### February 9, 2022 - [Diff](https://github.com/NREL/OpenStudio-BuildStock/compare/v2.4.0...v2.5.0)