+++ resources/hpxml-measures/Changelog.md
@@ -2,8 +2,10 @@
 
 __New Features__
 - **Breaking Change**: Replaces `FrameFloors/FrameFloor` with `Floors/Floor`.
+- Allows SEER2/HSPF2 efficiency types for central air conditioners and heat pumps.
 - Allows heating/cooling seasons that don't span the entire year.
 - Allows calculating one or more utility bill scenarios (e.g., net metering vs feed-in tariff compensation types for a simulation with PV).
+- Allows setting the EnergyPlus temperature capacitance multiplier.
 - EnergyPlus modeling changes:
   - Switches Kiva foundation model timestep from 'Hourly' to 'Timestep'; small increase in runtime for sub-hourly simulations.
 - Annual/timeseries outputs:
@@ -15,7 +17,7 @@ __New Features__
   - **Breaking change**: Replaces arguments using 'auto' for defaults with optional arguments of the appropriate data type. New `heat_pump_sizing_methodology` argument and new boolean `foo_present` arguments for lighting, appliances, etc.
   - Adds optional arguments for utility bill scenarios.
 - ReportUtilityBills measure:
-  - Removes utility rate and PV related arguments in lieu of new utility bill scenarios feature.
+  - Removes utility rate and PV related arguments in lieu of new utility bill scenarios described inside the HPXML file.
 
 __Bugfixes__
 - Fixes possible incorrect autosizing of heat pump *separate* backup systems with respect to duct loads.
@@ -26,6 +28,10 @@ __Bugfixes__
 - Fixes possible output error for ground source heat pumps with a shared hydronic circulation loop.
 - Provides an error message if the EnergyPlus simulation used infinite energy.
 - Fixes zero energy use for a ventilation fan w/ non-zero fan power and zero airflow rate.
+- Fixes excessive heat transfer when foundation wall interior insulation does not start from the top of the wall.
+- Fixes how relative paths are treated when using an OpenStudio Workflow.
+- BuildResidentialHPXML measure:
+  - Fixes aspect ratio convention for single-family attached and multifamily dwelling units.
 
 ## OpenStudio-HPXML v1.4.0